refactor(installations): split types into installations (public) and namespaced (deprecated)
